Extract a bit more sync reporting
        * Changes the name of directory `doChannelReport` to `reports`
        * Changes name of `attachments` to `convertStandups`
        * Extracts generation of field into reports directory
