+++ lib/helpers/doChannelReport.js
@@ -5,7 +5,8 @@ const async               = require('async');
 const _                   = require('underscore');
 const models              = require('../../models');
 const timeHelper          = require('./time');
-const createAttachments   = require('./doChannelReport/attachments');
+const convertStandups     = require('./reports/convertStandups');
+const generateFields      = require('./reports/generateFields');
 
 module.exports = function doChannelReport(bot, reportChannel, update, userRealName, actionCallback) {
   actionCallback = actionCallback || function () {};
@@ -47,91 +48,15 @@ module.exports = function doChannelReport(bot, reportChannel, update, userRealNa
         channel: channel.name
       };
 
-      var attachments = createAttachments(standups);
+      var attachments = convertStandups(standups);
+      attachments.unshift({
+        fallback: 'Todays standup for <#'+channel.name+'>',
+        pretext: 'Todays standup for <#'+channel.name+'>',
+        title: 'Summary',
+        fields: generateFields(attachments)
+      });
 
       async.series([
-        // syncronous operation, does not need to be inside async
-        function(callback) {
-          // Create summary statistics
-          var fields = [];
-
-          // Find common channels
-          var regex = /<#\w+>/g;
-          var search;
-          var results = {};
-          var commonChannels = '';
-          while ((search = regex.exec(JSON.stringify(attachments))) !== null) {
-            if (results[search[0]]) {
-              results[search[0]] += 1;
-            } else {
-              results[search[0]] = 1;
-            }
-          }
-          for (var i in results) {
-            if (results[i] > 1) {
-              // common[i] = results[i];
-              commonChannels += '- ' + i + ' ('+results[i]+')\n';
-            }
-          }
-
-          // Find people who used :pager: to indicate availability
-          // Find people who are out of office (OOO) today
-          var pager = '';
-          var ooo = '';
-          for (let attachment of attachments) {
-            for (let field of attachment.fields) {
-              if (field.title === 'Today') {
-                if (field.value.search(/:pager:/) >= 0) {
-                  pager += '- '+attachment.title + '\n';
-                }
-                if (field.value.toLowerCase() === 'ooo' || field.value.toLowerCase() === 'out of office') {
-                  ooo += '- '+attachment.title + '\n';
-                }
-              }
-            }
-          }
-
-          // Find total number of standups
-          var length = attachments.length;
-
-          // Assemble stats
-          fields.push({
-            title: 'Heard from',
-            value: length + ' people',
-            short: true
-          });
-          if (commonChannels.length >= 1) {
-            fields.push({
-              title: 'Common projects',
-              value: commonChannels,
-              short: false
-            });
-          }
-          if (pager.length >= 1) {
-            fields.push({
-              title: 'Available today :pager:',
-              value: pager,
-              short: false
-            });
-          }
-          if (ooo.length >= 1) {
-            fields.push({
-              title: 'Out today',
-              value: ooo,
-              short: false
-            });
-          }
-
-          // Add starter attachment
-          attachments.unshift({
-            fallback: 'Todays standup for <#'+channel.name+'>',
-            pretext: 'Todays standup for <#'+channel.name+'>',
-            title: 'Summary',
-            fields: fields
-          });
-          callback(null);
-        },
-
         // Send that report off to Slack
         function(callback) {
           if (!update) {
